locagent_kernel/requirements.txt
# Core dependencies
networkx>=2.8.0
pyyaml>=6.0.0

# LLM clients
openai>=1.0.0
anthropic>=0.3.0
litellm>=1.0.0

# Text processing and search
scikit-learn>=1.0.0  # For TF-IDF if needed as fallback

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0